Expand the event_mask field in controller conns to 64 bits
Back in 078d6bcd, we added an event number 0x20, but we didn't make
the event_mask field big enough to compensate.

Patch by "teor". Fixes 13085; bugfix on 0.2.5.1-alpha.

+++ src/or/or.h
@@ -1737,8 +1737,9 @@ typedef struct dir_connection_t {
 typedef struct control_connection_t {
   connection_t base_;
 
-  uint32_t event_mask; /**< Bitfield: which events does this controller
-                        * care about? */
+  uint64_t event_mask; /**< Bitfield: which events does this controller
+                        * care about?
+                        * EVENT_MAX_ is >31, so we need a 64 bit mask */
 
   /** True if we have sent a protocolinfo reply on this connection. */
   unsigned int have_sent_protocolinfo:1;
